refactor: remove Tailwind CSS components and update imports for consistency
