+++ packages/frontend/src/components/design/chess/train/HintInfo.tsx
@@ -1,11 +1,10 @@
-import { Box, TextField, Typography } from "@mui/material";
 import React from "react"
 import { MoveVariantNode } from "../../../../models/VariantNode";
+import { Textarea } from "@headlessui/react";
 
 interface HintInfoProps {
   currentMoveNode: MoveVariantNode;
 }
-
 
 export const HintInfo: React.FC<HintInfoProps> = ({
   currentMoveNode
@@ -33,19 +32,16 @@ export const HintInfo: React.FC<HintInfoProps> = ({
     }
    
     return (
-        <Box>
-          <Typography variant="h6" gutterBottom>
-            Comments
-          </Typography>
-      <TextField
-        label="No comments"
-        multiline
-        disabled={true}
-        rows={10}
-        value={getHints().join("\n")}
-        variant="outlined"
-        fullWidth
-      />
-        </Box>
-      );
+      <div className="p-4 bg-gray-800 rounded shadow-md flex flex-col h-full">
+        <h6 className="text-lg font-semibold mb-2">
+          Comments
+        </h6>
+        <Textarea
+          className="flex-grow p-2 bg-gray-700 text-white rounded focus:outline-none focus:ring-2 focus:ring-accent overflow-auto"
+          rows={10}
+          value={getHints().join("\n")}
+          disabled
+        ></Textarea>
+      </div>
+    );
 }
