+++ packages/frontend/src/components/design/chess/train/TrainInfo.tsx
@@ -1,20 +1,10 @@
-// TrainInfo.tsx
 import React, { useState } from "react";
-import {
-  Grid,
-  Typography,
-  Card,
-  CardContent,
-  Theme,
-  useMediaQuery,
-  LinearProgress,
-  Box,
-  Button,
-} from "@mui/material";
+import { Disclosure } from "@headlessui/react";
+import { ChevronUpIcon } from "@heroicons/react/20/solid";
 import whiteKing from "../../../../assets/white-king.svg";
 import blackKing from "../../../../assets/black-king.svg";
 import { getMovementsFromVariant } from "../../../../utils/chess/variants/getMovementsFromVariants";
-import { Turn } from "../../../../../../common/src/types/Orientation";
+import { Turn } from "@chess-opening-master/common/src/types/Orientation";
 import { TrainVariant } from "../../../../models/chess.models";
 import { MoveVariantNode } from "../../../../models/VariantNode";
 import { variantToPgn } from "../../../../utils/chess/pgn/pgn.utils";
@@ -37,10 +27,6 @@ const TrainInfo: React.FC<TrainInfoProps> = ({
   currentMoveNode
 }) => {
 
-  const isMobile = useMediaQuery((theme: Theme) =>
-    theme.breakpoints.down("sm")
-  );
-
   const currentVariant = trainVariants.filter(
     (variant) => variant.state === "finished"
   ).length;
@@ -59,126 +45,93 @@ const TrainInfo: React.FC<TrainInfoProps> = ({
   const handleCopyPgn = () => {
     if (lastTrainVariant) {
       const pgn = variantToPgn(lastTrainVariant.variant, turn, new Date());
-      const textArea = document.createElement("textarea");
-      textArea.value = pgn;
-      document.body.appendChild(textArea);
-      textArea.select();
-      document.execCommand("copy");
-      document.body.removeChild(textArea);
+      navigator.clipboard.writeText(pgn);
     }
   };
 
   return (
-    <Card>
-      <CardContent>
-        {lastTrainVariant && (
-          <Box display="flex" flexDirection="column" alignItems="flex-start" mb={2}>
-            <Typography variant="h6" gutterBottom>
-              Last Finished Variant
-            </Typography>
-            <Box display="flex" alignItems="center" justifyContent="space-between" width="100%">
-              <Typography variant="body1">{lastTrainVariant.variant.fullName}</Typography>
-              <Button variant="contained" color="primary" onClick={handleCopyPgn}>
-                Copy PGN
-              </Button>
-            </Box>
-          </Box>
+    <div className="shadow rounded-lg p-6 bg-gray-800">
+      {lastTrainVariant && (
+        <div className="mb-4">
+          <h2 className="text-lg font-semibold text-gray-200 dark:text-gray-300">Last Finished Variant</h2>
+          <div className="flex justify-between items-center mt-2">
+            <span className="text-md text-gray-400 dark:text-gray-500">{lastTrainVariant.variant.fullName}</span>
+            <button
+              onClick={handleCopyPgn}
+              className="bg-accent text-primary px-3 py-1 rounded hover:bg-yellow-500 transition-colors"
+            >
+              Copy PGN
+            </button>
+          </div>
+        </div>
+      )}
+      <div className="flex items-center mb-4">
+        {!finishedTrain && (
+          <img
+            src={turn === "white" ? whiteKing : blackKing}
+            alt={`${turn} king`}
+            className="w-8 h-8 mr-2"
+          />
         )}
-        <Grid container alignItems="center" spacing={2}>
-          {!finishedTrain && (
-            <Grid item>
-              {turn === "white" ? (
-                <img src={whiteKing} alt="white king" />
-              ) : (
-                <img src={blackKing} alt="black king" />
-              )}
-            </Grid>
-          )}
-          <Grid item>
-            <Typography variant="h5" component="div">
-              {finishedTrain
-                ? "Finished Training"
-                : isYourTurn
-                ? "Your turn"
-                : "Opponent's turn"}
-            </Typography>
-            {!isMobile && (
-              <Typography variant="subtitle1" color="textSecondary">
-                {finishedTrain
-                  ? "You have finished one full variant"
-                  : isYourTurn
-                  ? "Play one of your allowed moves according to your repertoire"
-                  : "Wait for your opponent to play"}
-              </Typography>
-            )}
-          </Grid>
-        </Grid>
-        <Typography variant="body1" style={{ marginTop: 16 }}>
-          {currentVariant} de {totalVariants} variantes
-        </Typography>
-        <LinearProgress
-          value={(currentVariant / totalVariants) * 100}
-          variant="determinate"
-          style={{ marginTop: 16, marginBottom: 16 }}
-        />
-        <Grid container spacing={2}>
-          <Grid item xs={12}>
-            <Typography variant="h6" style={{ marginBottom: 8 }}>
-              Available Variants to Play
-            </Typography>
-          </Grid>
-          <Grid item xs={12} style={{ maxHeight: '400px', overflowY: 'auto' }}>
-            <ul style={{ listStyleType: "none", padding: 0, margin: 0 }}>
-              {availableVariants.map((variant, index) => (
-                <li
-                  key={index}
-                  style={{
-                    display: "flex",
-                    flexDirection: "column",
-                    padding: "16px",
-                    marginBottom: "12px",
-                    backgroundColor: "#f9f9f9",
-                    borderRadius: "8px",
-                    border: "1px solid #ddd",
-                    boxShadow: "0px 4px 6px rgba(0, 0, 0, 0.1)",
-                    cursor: "pointer",
-                  }}
-                  onClick={() => handleToggleVariant(index)}
-                >
-                  <div style={{ display: "flex", alignItems: "center" }}>
-                    <span
-                      style={{
-                        width: "12px",
-                        height: "12px",
-                        borderRadius: "50%",
-                        backgroundColor: "#3f51b5",
-                        marginRight: "12px",
-                        flexShrink: 0,
-                      }}
-                    ></span>
-                    <Typography variant="body1" style={{ fontWeight: 500, flexGrow: 1 }}>
-                      {variant.variant.fullName}
-                    </Typography>
-                    <Typography variant="body2" color="textSecondary">
-                      Click to see variant
-                    </Typography>
-                  </div>
-                  {expandedVariant === index && (
-                    <div style={{ marginTop: "8px", paddingLeft: "24px", display: "flex", flexWrap: "wrap" }}>
+        <h3 className="text-xl font-bold text-gray-100 dark:text-gray-50">
+          {finishedTrain
+            ? "Finished Training"
+            : isYourTurn
+            ? "Your turn"
+            : "Opponent's turn"}
+        </h3>
+      </div>
+      {!finishedTrain && (
+        <p className="text-gray-400 dark:text-gray-300 mb-4">
+          {isYourTurn
+            ? "Play one of your allowed moves according to your repertoire."
+            : "Wait for your opponent to play."}
+        </p>
+      )}
+      <div className="mb-4">
+        <span className="text-sm text-gray-500 dark:text-gray-400">
+          {`${currentVariant} of ${totalVariants} variants`}
+        </span>
+        <div className="w-full bg-gray-700 rounded-full h-2.5 dark:bg-gray-600 mt-1">
+          <div
+            className="bg-accent h-2.5 rounded-full"
+            style={{ width: `${(currentVariant / totalVariants) * 100}%` }}
+          ></div>
+        </div>
+      </div>
+      <div>
+        <h4 className="text-lg font-semibold text-gray-200 dark:text-gray-300 mb-2">
+          Available Variants to Play
+        </h4>
+        <ul className="space-y-2">
+          {availableVariants.map((variant, index) => (
+            <Disclosure key={index}>
+              {({ open }) => (
+                <>
+                  <Disclosure.Button className="flex justify-between w-full px-4 py-2 text-left text-sm font-medium text-gray-100 dark:text-gray-200 bg-gray-700 dark:bg-gray-600 rounded-lg hover:bg-gray-600 dark:hover:bg-gray-500 focus:outline-none focus-visible:ring focus-visible:ring-purple-500 focus-visible:ring-opacity-75">
+                    <span>{variant.variant.fullName}</span>
+                    <ChevronUpIcon
+                      className={`${
+                        open ? "transform rotate-180" : ""
+                      } w-5 h-5 text-accent`}
+                    />
+                  </Disclosure.Button>
+                  <Disclosure.Panel className="px-4 pt-4 pb-2 text-sm text-gray-400 dark:text-gray-300">
+                    <div className="flex flex-wrap gap-2">
                       {getMovementsFromVariant(variant, currentMoveNode).map((move, moveIndex) => (
-                        <Typography key={moveIndex} variant="body2" style={{ marginRight: "16px", marginBottom: "8px" }}>
+                        <span key={moveIndex} className="px-2 py-1 bg-gray-200 dark:bg-gray-600 text-gray-700 dark:text-gray-200 rounded">
                           {move}
-                        </Typography>
+                        </span>
                       ))}
                     </div>
-                  )}
-                </li>
-              ))}
-            </ul>
-          </Grid>
-        </Grid>
-      </CardContent>
-    </Card>
+                  </Disclosure.Panel>
+                </>
+              )}
+            </Disclosure>
+          ))}
+        </ul>
+      </div>
+    </div>
   );
 };
