+++ packages/frontend/src/components/design/layouts/MainContainer.tsx
@@ -1,18 +1,9 @@
 import React from "react";
 
-const contentStyles: React.CSSProperties = {
-    flexGrow: 1,
-    padding: "1rem",
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-    justifyContent: "center",
-  };
-
-  export const MainContainer = ({ children }: { children: React.ReactNode }) => {
-    return (
-      <main style={contentStyles}>
-        {children}
-      </main>
-    );
-  }
+export const MainContainer = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <main className="flex-grow p-4 flex flex-col items-center justify-center scrollbar-custom bg-background h-full">
+      {children}
+    </main>
+  );
+};
