--- packages/frontend/src/components/design/lichess/ResultBar.css
@@ -1,33 +1,0 @@
-.bar {
-  display: flex;
-  width: 100%;
-  height: 16px;
-  border-radius: 8px;
-  overflow: hidden;
-  box-shadow: 0 0 4px rgba(0, 0, 0, 0.2);
-}
-
-.bar span {
-  text-align: center;
-  display: inline-block;
-  overflow: hidden;
-  vertical-align: middle;
-  line-height: 16px;
-  font-size: 0.8em;
-  transition: width 1s;
-}
-
-.bar .white {
-  background: linear-gradient(to bottom, #e6e6e6, #ccc);
-  color: black;
-}
-
-.bar .draws {
-  background: linear-gradient(to bottom, #777, #555);
-  color: white;
-}
-
-.bar .black {
-  background: linear-gradient(to bottom, #222, black);
-  color: white;
-}
