+++ packages/frontend/src/components/design/lichess/ResultBar.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import './ResultBar.css';
 
 interface ResultBarProps {
   winPercentage: number;
@@ -9,14 +8,14 @@ interface ResultBarProps {
 
 const ResultBar: React.FC<ResultBarProps> = ({ winPercentage, drawPercentage, lossPercentage }) => {
   return (
-    <div className="bar">
-      <span className="white" style={{ width: `${winPercentage}%` }}>
+    <div className="flex w-full h-4 rounded overflow-hidden shadow-sm">
+      <span className="flex justify-center items-center text-xs transition-width duration-1000 bg-gray-300 text-black" style={{ width: `${winPercentage}%` }}>
         {winPercentage.toFixed(1)}%
       </span>
-      <span className="draws" style={{ width: `${drawPercentage}%` }}>
+      <span className="flex justify-center items-center text-xs transition-width duration-1000 bg-gray-500 text-white" style={{ width: `${drawPercentage}%` }}>
         {drawPercentage.toFixed(1)}%
       </span>
-      <span className="black" style={{ width: `${lossPercentage}%` }}>
+      <span className="flex justify-center items-center text-xs transition-width duration-1000 bg-black text-white" style={{ width: `${lossPercentage}%` }}>
         {lossPercentage.toFixed(1)}%
       </span>
     </div>
