+++ packages/frontend/src/pages/repertoires/CreateRepertoire.tsx
@@ -1,23 +1,7 @@
-import {
-  Button,
-  Container,
-  Grid,
-  Typography,
-  TextField,
-  styled,
-} from "@mui/material";
 import React, { useEffect, useState } from "react";
 import { createRepertoire } from "../../repository/repertoires/repertoires";
 import { useNavbarContext } from "../../contexts/NavbarContext";
 import { useNavigate } from "react-router-dom";
-
-const CreateRepertoireContent = styled("div")`
-  padding: ${(props) => props.theme.spacing(8, 0, 6)};
-`;
-
-const CreateRepertoireButtons = styled("div")`
-  margin-top: ${(props) => props.theme.spacing(4)};
-`;
 
 const CreateRepertoire: React.FC = () => {
   const [repertoireName, setRepertoireName] = useState("");
@@ -35,47 +19,30 @@ const CreateRepertoire: React.FC = () => {
     const createdRepertoire = await createRepertoire(repertoireName);
     const id = createdRepertoire.insertedId;
     updateRepertoires();
-    navigate(`/repertoire/${id}`)
-    
+    navigate(`/repertoire/${id}`);
   };
 
   return (
-    <div>
-      <CreateRepertoireContent>
-        <Container maxWidth="sm">
-          <Typography component="h1" variant="h2" align="center" gutterBottom>
-            Crear nuevo repertorio
-          </Typography>
-          <Typography variant="h5" align="center" paragraph>
-            Ingresa el nombre de tu nuevo repertorio de aperturas personalizado.
-          </Typography>
-          <TextField
-            variant="outlined"
-            margin="normal"
-            required
-            fullWidth
-            id="repertoireName"
-            label="Nombre del repertorio"
-            name="repertoireName"
-            autoFocus
-            value={repertoireName}
-            onChange={handleChange}
-          />
-          <CreateRepertoireButtons>
-            <Grid container spacing={2} justifyContent="center">
-              <Grid item>
-                <Button
-                  variant="contained"
-                  color="primary"
-                  onClick={handleCreateRepertoire}
-                >
-                  Crear repertorio
-                </Button>
-              </Grid>
-            </Grid>
-          </CreateRepertoireButtons>
-        </Container>
-      </CreateRepertoireContent>
+    <div className="flex flex-col items-center justify-center min-h-screen bg-background text-textLight">
+      <div className="max-w-md w-full p-8 bg-gray-800 rounded shadow-md">
+        <h1 className="text-2xl font-bold text-center mb-4">Crear nuevo repertorio</h1>
+        <p className="text-center mb-4">Ingresa el nombre de tu nuevo repertorio de aperturas personalizado.</p>
+        <input
+          type="text"
+          className="w-full p-2 mb-4 bg-gray-700 text-white rounded focus:outline-none focus:ring-2 focus:ring-accent"
+          placeholder="Nombre del repertorio"
+          value={repertoireName}
+          onChange={handleChange}
+        />
+        <div className="flex justify-center">
+          <button
+            className="px-4 py-2 bg-accent text-black font-medium rounded hover:bg-yellow-500"
+            onClick={handleCreateRepertoire}
+          >
+            Crear repertorio
+          </button>
+        </div>
+      </div>
     </div>
   );
 };
